move input to separate component
